+++ src/wrap_trigger_shutdown.rs
@@ -2,16 +2,16 @@ use std::future::Future;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
-use crate::VitalToken;
+use crate::TriggerShutdownToken;
 
-/// Wrapped future that triggers a shutdown when it completes or is dropped.
+/// Wrapped future that triggers a shutdown when it completes or when it is dropped.
 #[must_use = "futures must be polled to make progress"]
-pub struct WrapVital<F> {
-	pub(crate) vital_token: Option<VitalToken>,
+pub struct WrapTriggerShutdown<T: Clone, F> {
+	pub(crate) trigger_shutdown_token: Option<TriggerShutdownToken<T>>,
 	pub(crate) future: F,
 }
 
-impl<F: Future> Future for WrapVital<F> {
+impl<T: Clone, F: Future> Future for WrapTriggerShutdown<T, F> {
 	type Output = F::Output;
 
 	#[inline]
@@ -22,7 +22,7 @@ impl<F: Future> Future for WrapVital<F> {
 			match Pin::new_unchecked(&mut me.future).poll(context) {
 				Poll::Pending => Poll::Pending,
 				Poll::Ready(value) => {
-					me.vital_token = None;
+					me.trigger_shutdown_token = None;
 					Poll::Ready(value)
 				},
 			}