+++ tests/shutdown.rs
@@ -3,7 +3,7 @@ use futures::future;
 use std::future::Future;
 use std::time::Duration;
 
-use async_shutdown::Shutdown;
+use async_shutdown::ShutdownManager;
 
 #[track_caller]
 fn test_timeout(test: impl Future<Output = ()>) {
@@ -16,48 +16,61 @@ fn test_timeout(test: impl Future<Output = ()>) {
 
 #[test]
 fn shutdown() {
-	// Create a `Shutdown` manager and instantly trigger a shutdown,
+	// Create a `ShutdownManager` manager and instantly trigger a shutdown,
 	test_timeout(async {
-		let shutdown = Shutdown::new();
-		shutdown.shutdown();
-		shutdown.wait_shutdown_triggered().await;
-		shutdown.wait_shutdown_complete().await;
+		let shutdown = ShutdownManager::new();
+		assert!(let Ok(()) = shutdown.trigger_shutdown(1));
+		assert!(shutdown.wait_shutdown_triggered().await == 1);
+		assert!(shutdown.wait_shutdown_complete().await == 1);
 	});
 
 	// Trigger the shutdown from a task after a short sleep.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 
 		tokio::spawn({
 			let shutdown = shutdown.clone();
 			async move {
 				tokio::time::sleep(Duration::from_millis(20)).await;
-				shutdown.shutdown();
+				assert!(let Ok(()) = shutdown.trigger_shutdown(2));
 			}
 		});
 
-		shutdown.wait_shutdown_triggered().await;
-		shutdown.wait_shutdown_complete().await;
+		assert!(shutdown.wait_shutdown_triggered().await == 2);
+		assert!(shutdown.wait_shutdown_complete().await == 2);
 	});
+}
+
+#[test]
+fn shutdown_only_works_once() {
+	let shutdown = ShutdownManager::new();
+	assert!(let Ok(()) = shutdown.trigger_shutdown("first"));
+	assert!(let Err(async_shutdown::ShutdownAlreadyStarted {
+		shutdown_reason: "first",
+		ignored_reason: "second",
+		..
+	}) = shutdown.trigger_shutdown("second"));
 }
 
 #[test]
 fn wrap_cancel() {
 	// Spawn a never-completing future that is automatically dropped on shutdown.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 		let task = tokio::spawn(shutdown.wrap_cancel(future::pending::<()>()));
-		shutdown.shutdown();
-		assert!(let Ok(None) = task.await);
+		assert!(let Ok(()) = shutdown.trigger_shutdown("goodbye!"));
+		let_assert!(Ok(Err(reason)) = task.await);
+		assert!(reason == "goodbye!");
 	});
 
 	// Same test, but use a ShutdownSignal to wrap the future.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 		let wait_shutdown_triggered = shutdown.wait_shutdown_triggered();
 		let task = tokio::spawn(wait_shutdown_triggered.wrap_cancel(future::pending::<()>()));
-		shutdown.shutdown();
-		assert!(let Ok(None) = task.await);
+		assert!(let Ok(()) = shutdown.trigger_shutdown("wooh"));
+		let_assert!(Ok(Err(reason)) = task.await);
+		assert!(reason == "wooh");
 	});
 }
 
@@ -65,19 +78,19 @@ fn wrap_cancel() {
 fn wrap_cancel_no_shutdown() {
 	// Spawn an already ready future and verify that it can complete if no shutdown happens.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::<()>::new();
 		let task = tokio::spawn(shutdown.wrap_cancel(future::ready(10)));
-		assert!(let Ok(Some(10)) = task.await);
+		assert!(let Ok(Ok(10)) = task.await);
 	});
 
 	// Same test, but use a future that first sleeps and then completes.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::<()>::new();
 		let task = tokio::spawn(shutdown.wrap_cancel(async {
 			tokio::time::sleep(Duration::from_millis(20)).await;
 			10u32
 		}));
-		assert!(let Ok(Some(10)) = task.await);
+		assert!(let Ok(Ok(10)) = task.await);
 	});
 }
 
@@ -85,13 +98,14 @@ fn wrap_cancel_no_shutdown() {
 fn delay_token() {
 	// Delay shutdown completion using a delay-shutdown token.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 
 		// Create a token to delay shutdown completion.
 		let_assert!(Ok(delay) = shutdown.delay_shutdown_token());
 
-		shutdown.shutdown();
+		assert!(let Ok(()) = shutdown.trigger_shutdown(10));
 		shutdown.wait_shutdown_triggered().await;
+		assert!(shutdown.is_shutdown_completed() == false);
 
 		// Move the token into a task where it is dropped after a short sleep.
 		tokio::spawn(async move {
@@ -107,11 +121,11 @@ fn delay_token() {
 fn wrap_delay() {
 	// Spawn a future that delays shutdown as long as it is running.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 		let_assert!(Ok(delay) = shutdown.delay_shutdown_token());
-		shutdown.shutdown();
+		assert!(let Ok(()) = shutdown.trigger_shutdown(10));
 
-		tokio::spawn(delay.wrap_wait(async move {
+		tokio::spawn(delay.wrap_future(async move {
 			tokio::time::sleep(Duration::from_millis(10)).await;
 		}));
 
@@ -123,37 +137,37 @@ fn wrap_delay() {
 #[test]
 fn delay_token_too_late() {
 	// Try go get a delay-shutdown token after the shutdown completed.
-	let shutdown = Shutdown::new();
-	shutdown.shutdown();
+	let shutdown = ShutdownManager::new();
+	assert!(let Ok(()) = shutdown.trigger_shutdown(()));
 	assert!(let Err(async_shutdown::ShutdownAlreadyCompleted { .. }) = shutdown.delay_shutdown_token());
-	assert!(let Err(async_shutdown::ShutdownAlreadyCompleted { .. }) = shutdown.wrap_wait(future::pending::<()>()));
+	assert!(let Err(async_shutdown::ShutdownAlreadyCompleted { .. }) = shutdown.wrap_delay_shutdown(future::pending::<()>()));
 }
 
 #[test]
 fn vital_token() {
-	// Trigger a shutdown by dropping a vital token.
+	// Trigger a shutdown by dropping a token.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 
-		let vital = shutdown.vital_token();
-		drop(vital);
+		let trigger_shutdown_token = shutdown.trigger_shutdown_token("stop!");
+		drop(trigger_shutdown_token);
 
-		shutdown.wait_shutdown_triggered().await;
-		shutdown.wait_shutdown_complete().await;
+		assert!(shutdown.wait_shutdown_triggered().await == "stop!");
+		assert!(shutdown.wait_shutdown_complete().await == "stop!");
 	});
 
 	// Same test, but drop the vital token in a task.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 
-		let vital = shutdown.vital_token();
+		let vital = shutdown.trigger_shutdown_token("done");
 		tokio::spawn(async move {
 			tokio::time::sleep(Duration::from_millis(20)).await;
 			drop(vital);
 		});
 
-		shutdown.wait_shutdown_triggered().await;
-		shutdown.wait_shutdown_complete().await;
+		assert!(shutdown.wait_shutdown_triggered().await == "done");
+		assert!(shutdown.wait_shutdown_complete().await == "done");
 	});
 }
 
@@ -161,24 +175,24 @@ fn vital_token() {
 fn wrap_vital() {
 	// Trigger a shutdown by dropping a vital token from a task after a short sleep.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 
-		tokio::spawn(shutdown.wrap_vital(async move {
+		tokio::spawn(shutdown.wrap_trigger_shutdown("sleep done", async move {
 			tokio::time::sleep(Duration::from_millis(20)).await;
 		}));
 
-		shutdown.wait_shutdown_triggered().await;
-		shutdown.wait_shutdown_complete().await;
+		assert!(shutdown.wait_shutdown_triggered().await == "sleep done");
+		assert!(shutdown.wait_shutdown_complete().await == "sleep done");
 	});
 
 	// Same test, but now use a future that is instantly ready.
 	test_timeout(async {
-		let shutdown = Shutdown::new();
+		let shutdown = ShutdownManager::new();
 
 		// Trigger the shutdown by dropping a vital token from a task after a short sleep.
-		tokio::spawn(shutdown.wrap_vital(future::ready(())));
+		tokio::spawn(shutdown.wrap_trigger_shutdown("stop", future::ready(())));
 
-		shutdown.wait_shutdown_triggered().await;
-		shutdown.wait_shutdown_complete().await;
+		assert!(shutdown.wait_shutdown_triggered().await == "stop");
+		assert!(shutdown.wait_shutdown_complete().await == "stop");
 	});
 }