+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# Unreleased:
+* Rename `Shutdown` struct to `ShutdownManager`.
+* Rename `ShutdownManager` methods:
+  * `shutdown()` is now called `trigger_shutdown()`
+  * `shutdown_started()` is now called `is_shutdown_triggered()`.
+  * `shutdown_completed()` is now called `is_shutdown_completed()`.
+  * `wrap_vital()` is now called `wrap_trigger_shutdown()`.
+  * `wrap_wait()` is now called `wrap_delay_shutdown()`.
+* Rename `VitalToken` to `TriggerShutdownToken`.
+* Rename `TriggerShutdownToken::wrap_vital()` to `wrap_future()`.
+* Rename `DelayShutdownToken::wrap_wait()` to `wrap_future()`.
+* All types now take a generic parameter `T: Clone` for the shutdown reason.
+* Add a parameter for the shutdown reason in `ShutdownManager` methods `trigger_shutdown()`, `wrap_trigger_shutdown()` and `trigger_shutdown_token()`.
+* Add `ShutdownManager::shutdown_reason()` to retrieve the shutdown reason.
+* Return the shutdown reason from `ShutdownManager::wait_shutdown_triggered()` and `ShutdownManager::wait_shutdown_complete()`.
+* Add `shutdown_reason` field to `ShutdownAlreadyCompleted` struct.
+* Return a `ShutdownAlreadyStarted` error when calling `trigger_shutdown()` multiple times.
+* Change the output type of `WrapCancel` futures from `Option` to a `Result` with the shutdown reason as error.
+
 # Version 0.1.3 - 2023-08-14
 * Mark all future wrappers as `#[must_use]`.
 
@@ -14,7 +33,7 @@
 * Regenerate README.md from library documentation.
 
 # Version 0.1.0-beta1 - 2021-10-04
-* Change `DelayShutdownToken::wrap_wait()` to constume the token.
+* Change `DelayShutdownToken::wrap_wait()` to consume the token.
 * Do not consume the `Shutdown` object in `Shutdown::wrap_vital()`.
 * Rename `wait_shutdown()` to `wait_shutdown_triggered()`.